Implement basic flash-cards-game functionality
Improve error handling by adding general exception handler

+++ src/main/java/mate/lingua/service/LearningDatasetServiceImpl.java
@@ -30,7 +30,11 @@ public class LearningDatasetServiceImpl implements LearningDatasetService {
     }
 
     @Override
-    public void deleteById(Long learningDatasetId) {
-        learningDatasetRepository.deleteById(learningDatasetId);
+    public boolean deleteById(Long learningDatasetId) {
+        if (learningDatasetRepository.existsById(learningDatasetId)) {
+            learningDatasetRepository.deleteById(learningDatasetId);
+            return true;
+        }
+        return false;
     }
 }
